fix() corrected to plane.fromVec3() and added tests
