+++ src/core/math/plane/fromPoints.js
@@ -1,5 +1,5 @@
 const vec3 = require('../../math/vec3')
-const fromVector3Ds = require('./fromVector3Ds')
+const fromVec3s = require('./fromVec3s')
 
 /**
  * Create a new plane from the given points
@@ -12,7 +12,7 @@ const fromPoints = (a, b, c) => {
   const v1 = vec3.fromVarious(a)
   const v2 = vec3.fromVarious(b)
   const v3 = vec3.fromVarious(c)
-  return fromVector3Ds(v1, v2, v3)
+  return fromVec3s(v1, v2, v3)
 }
 
 module.exports = fromPoints
