+++ src/core/math/plane/fromVec3s.js
@@ -3,12 +3,13 @@ const fromValues = require('../vec4/fromValues')
 
 /**
  * Create a new plane from the given vec3 values
+ *
  * @param {a} vec3 - vector 3D
  * @param {b} vec3 - vector 3D
  * @param {c} vec3 - vector 3D
  * @returns {Array} a new plane with properly typed values
  */
-const fromVector3Ds = (a, b, c) => {
+const fromVec3s = (a, b, c) => {
   // let n = b.minus(a).cross(c.minus(a)).unit()
   // FIXME optimize later
   const ba = vec3.subtract(b, a)
@@ -19,4 +20,4 @@ const fromVector3Ds = (a, b, c) => {
   return fromValues(normal[0], normal[1], normal[2], w)
 }
 
-module.exports = fromVector3Ds
+module.exports = fromVec3s
