+++ src/core/math/plane/index.js
@@ -5,9 +5,8 @@ module.exports = {
   flip: require('./flip'),
   fromNormalAndPoint: require('./fromNormalAndPoint'),
   fromObject: require('./fromObject'),
-  fromPoints: require('./fromPoints'),
   fromValues: require('../vec4/fromValues'),
-  fromVector3Ds: require('./fromVector3Ds'),
+  fromVec3s: require('./fromVec3s'),
   signedDistanceToPoint: require('./signedDistanceToPoint'),
   toString: require('../vec4/toString'),
   transformMat4: require('./transformMat4')
